requirements.txt
Django>=5
djangorestframework>=3.14
psycopg2-binary>=2.9
gunicorn

